third_party/blink/web_tests/svg/dynamic-updates/SVGFEConvolveMatrixElement-svgdom-kernelMatrix-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'kernelMatrix' property of the SVGFEConvolveMatrixElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS SVGNumberListToString(convolveMatrixElement.kernelMatrix.baseVal) is "-2 0 0 0 1 0 0 0 2 "
PASS SVGNumberListToString(convolveMatrixElement.kernelMatrix.baseVal) is "3 0 0 0 1 0 0 0 2 "
PASS successfullyParsed is true

TEST COMPLETE
